Moved Classes to a separate file (module.py)
Had to make some changes to make sure everything still interacted the same.
